refactor : camelCase에서 kebab-case로 리소스 폴더명 변경
