+++ src/file/profile-picture/dto/profile-picture-response.dto.ts
