+++ src/file/profile-picture/entities/profile-picture-type.enum.ts
