+++ src/file/profile-picture/entities/profile-picture.entity.ts
