+++ src/file/profile-picture/profile-picture.module.ts
